Move parsing logic into separate module to `parse_watched_topics_meta`
