+++ crates/torrent-bot-clients/src/toloka.rs
@@ -4,38 +4,10 @@ use scraper::{Html, Selector};
 use serde::Serialize;
 use tracing::warn;
 
-#[derive(Debug)]
-pub enum Category {
-    Movies,
-    Series,
-    Other(String),
-}
+use crate::toloka::types::{DownloadMeta, Topic, TopicMeta};
 
-impl ToString for Category {
-    fn to_string(&self) -> String {
-        String::from(match self {
-            Self::Movies => "Movies",
-            Self::Series => "Series",
-            Self::Other(_) => "Other",
-        })
-    }
-}
-
-pub struct TopicMeta {
-    pub topic_id: String,
-    pub title: String,
-    pub category: Category,
-}
-
-pub struct DownloadMeta {
-    pub registered_at: String,
-    pub download_id: String,
-}
-
-pub struct Topic {
-    pub topic_meta: TopicMeta,
-    pub download_meta: DownloadMeta,
-}
+pub(crate) mod parsers;
+pub mod types;
 
 const TOLOKA_HOST: &str = "https://toloka.to";
 
@@ -121,47 +93,9 @@ impl TolokaClient {
         }
 
         let document = response.text().await?;
-        let html = Html::parse_document(&document);
+        let topics_meta = parsers::parse_watched_topics_meta(&document);
 
-        let table_row_selector = Selector::parse(r#"table.forumline tr"#).unwrap();
-        let table_entries = html.select(&table_row_selector);
-
-        let href_selector = &Selector::parse(r#"a[href]"#).unwrap();
-        let td_selector = &Selector::parse(r#"td"#).unwrap();
-
-        let mut topics = vec![];
-
-        for el in table_entries
-            .skip(1)
-            .filter(|el| el.children().filter(|el| el.value().is_element()).count() == 6)
-            .into_iter()
-        {
-            let columns = el.select(&td_selector).collect::<Vec<_>>();
-            let link = columns[0].select(&href_selector).next().unwrap();
-
-            let category_raw = columns[1]
-                .select(&href_selector)
-                .next()
-                .unwrap()
-                .inner_html()
-                .to_string();
-            let topic_id = link.value().attr("href").unwrap_or_default().to_string();
-            let title = link.inner_html().to_string();
-
-            let category = match category_raw.to_lowercase().as_str() {
-                s if s.contains("фільм") => Category::Movies,
-                s if s.contains("серіал") => Category::Series,
-                other => Category::Other(other.to_string()),
-            };
-
-            topics.push(TopicMeta {
-                topic_id,
-                category,
-                title,
-            });
-        }
-
-        Ok(topics)
+        Ok(topics_meta)
     }
 
     async fn get_download_meta(&self, topic_id: &str) -> TolokaClientResult<Option<DownloadMeta>> {
